Practice/Practice 2/requirement.txt
polars == 0.20.10
numpy == 1.26.4
matplotlib == 3.8.2
varname == 0.13.0
torch == 2.1.1
lightning == 2.2.0.post0
torchinfo == 1.8.0
torchmetrics == 1.3.1
tensorboard == 2.16.2
pillow == 10.2.0
tqdm == 4.66.2
directory-tree == 0.0.4
ultralytics == 8.1.29
ipywidgets == 8.1.2



